plots: render single page by-default and introduce --show-json
