+++ tests/unit/command/test_plots.py
@@ -15,8 +15,7 @@ def test_metrics_diff(dvc, mocker):
             "datafile",
             "--select",
             "column1,column2",
-            "--no-html",
-            "--stdout",
+            "--show-json",
             "-x",
             "x_field",
             "-y",
@@ -46,7 +45,6 @@ def test_metrics_diff(dvc, mocker):
         revs=["HEAD", "tag1", "tag2"],
         fields={"column1", "column2"},
         path=None,
-        embed=False,
         x_field="x_field",
         y_field="y_field",
         csv_header=True,
@@ -67,8 +65,7 @@ def test_metrics_show(dvc, mocker):
             "template",
             "-s",
             "$.data",
-            "--no-html",
-            "--stdout",
+            "--show-json",
             "--no-csv-header",
             "datafile",
         ]
@@ -87,7 +84,6 @@ def test_metrics_show(dvc, mocker):
         template="template",
         fields=None,
         path="$.data",
-        embed=False,
         x_field=None,
         y_field=None,
         csv_header=False,
